CA-10404: xen-bugtool no longer collects the right databases CA-10406: xen-bugtool could collect crash dumps CP-14: Improve ability to collect xen-bugtool output
Use new xapi-db-process to retrieve the database, rather than grabbing it
from /var/xapi.  This command will parse db.conf correctly.  We still grab
local.db from /var/xapi, if present.

Add an s-expression parser, and use that to remove the cifspassword from the
database.  The other_config entries are all stored as s-expressions.  This
uses sqlite3 to edit the database directly (once copied using xapi-db-process).

Collect host crashdumps from /var/crash.  These are in two categories: one for
*.log, which are taken to be free of privacy problems and small, and one for
the rest of the files (i.e. *dump) which are assumed to potentially have
passwords and so on, and are large.

Recalculate the sizes for crashdumps and xenserver logs, so that the
capabilities are precise.

Add an md5sum to each entry in the inventory.

Rename all XENSOURCE entries to XENSERVER.