CP-9754: Capture /var/log/sa logs
There was some concern that these logs are too large to be captured.
However, since only 7 days of logs are kept and the logs compress really
well, I don't think this is a problem.

I have tested a host with about 300 VBDs and 400 vifs, forced sar to
capture the equivalent of 7 days worth of logs, and it results in the
bugtool being 3.4MB larger.  This is fairly small compared with the
overall size of the bugtool.  On most hosts, this figure would probably
be even smaller.