Only request RRDs for VMs which are resident in the pool
This issue was present forever, so it's very minor it but can be confusing:

To reproduce, run this command on a Pool master:

root@master# xe vm-install template="Oracle Linux 8" new-name-label=empty-VM
root@master# xe vm-list params=resident-on name-label=empty-VM
Output:
resident-on ( RO)    : <not in database>

xen-bugtool's dump_xapi_rrds() would happily request RRDs for this VM,
but would trigger an error in xapi-rrdd and get an error response.

To avoid this, don't request RRDs for such VM records from xapi-rrdd.

Uses "vm.get('resident_on', None)" to ensure that this never triggers
an exception.