CP-45506: Archive /etc/systemd as a tar archive inside the output
xenserver-status-report also collects the files and symlinks below
/etc/systemd. The paths within in can become very long, especially
in the presence of "templated" services such as those used by the
GFS2 SR which bind to the file system mount point and get created
with escaped uuids in their names.

This subsequently causes issues for XenCenter as the version of .NET
being used (at least in the CH 8.2 CU1 version) has a very short
maximum path limit and fails when repacking the status report.

To mitigate against this for CU1 (XenCenter for XenServer 8 will be
fixed to handle longer paths) update the bugtool to package the contents
of /etc/systemd as a tarball instead of loose files and symlinks:

TarOutput and ZipOutput are now subclasses of ArchiveWithTarSubarchives
which implements declare_subarchive(define the paths of subarchives),
add_path_to_subarchive(add a file to a defined subarchive) and
add_subarchives(close subarchives and add them to the output).

The latter is called from ZipOutput/TarOutput's close() method
to finalize the subarchives before writing the output archives.

add_path_to_subarchive() is called by by ZipOutput/TarOutput's
addRealFile() to add the file to the subarchive if needed instead.

declare_subarchive() is called by main() to define the paths
which shall be captured into the respective subarchives.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>