Fix getting output from `sar -A`, moving sar to "system-load"
Fix `sar -A` from timing out:

This issue was that when the human-readable output of `sar -A`
was added in April to `--entries=xenserver-logs`, the timeout
for the capability was not defined.

The capability `xenserver-logs` uses the default timeout of -1
which is racey for commands as they are immediately stopped
due to timeout of -1 running out unless they are very brief.

To fix this, it was suggested that `sar` (a system load report)
could be moved to a new capability. Also, it should be noted that
`sar` records are binary system load records, not XenServer specific
log files.

Therefore, create a new capability (`system_load`), responsible for
collecting the sar data files (and the human-readable of output
of `sar -A`) with a timeout of 30 seconds and a `max_size` of 70 MB.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>