Update urlopen() and getoutput() to support Python3 as well
There are many updates ahead until it starts to work on Python3, and for
now, let's not cut off Python2 because when we do so, we cannot merge
that code to master as it not run on Python2 nor on Python3. To avoid
this, this simple compatibility handling allows us to merge to master.

Use line-length=108 for formatting by black and isort for fewer reformats

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>