CI: Improve unit tests to test the bugtool's main(): Coverage: 81%!
Add tests/unit/test_main.py which also uses test functions which were
added for earlier tests to test the main function of bugtool to produce
the bugtool output archives for all types: zip, tar, and tar.bz2.

The main test checks the output of the bugtool application to ensure that
it matches the expected output. It compares the captured output with the
expected output and performs various assertions to validate the output.

It extracts the output files from the archive and checks that the xAPI db
and the inventory.xml.

Tested output:
- The start and end of the collected bugtool messages from captured.out
- The inventory.xml file is checked to be validated using its XML schema.
- The xapi-db.xml is checked to have secrets filtered using a dummy xapi db

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>