CP-45506: Fix asking if /etc/systemd/* should be collected
For CP-45506 (Package /etc/systemd into a tarball), I had assumed
that I'd have to move the tree_output() for /etc/systemd to after
where the tarball for /etc/systemd is declared, but this was not
needed.

Moving it had the side effect that in interactive mode (a question
if each individual file shall be collected - hundreds of questions
usually), the /etc/systemd tarball was simply always collected.

This commit fixes this by reverting this move, back to the original
location of this line, to before the question if the files in
/etc/systemd/* shall be collected.

Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>