CP-46759: Fixup Test case on testsing the SNMP functions in bugtool.
TODO: The actual collection and filtering of files is not tested yet!

The unit test checks filtering the inputs using the individual funcs.

But:

With this commit, from an actual bugtool run,
we only test the non-existing files case!

Preferaby:

You should add the new config files as they are used now
to tests/integration/dom0-template, then verify that the
contents are the filtered contents.

However, without that final TODO done, we can already
merge this and do the actual collection of the files
in a new PR.

Co-authored-by: Bengang Yuan <bengang.yuan@cloud.com>
Signed-off-by: Bernhard Kaindl <bernhard.kaindl@cloud.com>