CA-392459: Avoid opening /dev/mem when calling biosdevname
biosdevname opens /dev/mem to read the $PIR PCI interrupt routing table.
This fails with Secure Boot enabled and causes a warning in the kernel.
log. Since the $PIR PCI interrupt routing table is used for routing PCI
interrupts to ISA IRQs, it is not useful on any modern system so pass
"-x" to biosdevname to avoid this behaviour.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>