CA-403717: Make XEN_RT mode additive for entries
The environment variable `XEN_RT` is used to specify that logs are being
collected by an automation tool, which wants *all* data, including that which
would not normally be collected / offered.

The previous code was hardcoding a set of entries, which meant any new entries
added e.g. by a bugtool plugin, were then not being captured, even if they
should have been included by default.

Fix this by making the code additive, so we simply add any of our predefined
entries which are missing from the default set.

Signed-off-by: Alex Brett <alex.brett@cloud.com>